+++ src/menu-item-recipe/menu-item-recipe.controller.ts
@@ -7,11 +7,14 @@ import {
   Body,
   Param,
   UsePipes,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { MenuItemRecipeService } from './menu-item-recipe.service';
 import { UpdateDateColumn } from 'typeorm';
 import { MenuItemRecipeDTO } from './menu-item-recipe.dto';
 import { ValidationPipe } from '../shared/validation.pipe';
+var validate = require('uuid-validate');
 
 @Controller('menu-item-recipe')
 export class MenuItemRecipeController {
@@ -24,6 +27,10 @@ export class MenuItemRecipeController {
 
   @Get(':menuItem_id')
   getMenuItemRecipeById(@Param('menuItem_id') menuItem_id: string) {
+    //Check id is valid uuid
+    if (!validate(menuItem_id)) {
+      throw new HttpException('invalid id format!', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemRecipeService.readById({ id: menuItem_id });
   }
 
@@ -41,6 +48,10 @@ export class MenuItemRecipeController {
 
   @Delete(':id')
   deleteMenuItemRecipe(@Param('id') id: string) {
+    //Check id is valid uuid
+    if (!validate(id)) {
+      throw new HttpException('invalid id format!', HttpStatus.BAD_REQUEST);
+    }
     return this.menuItemRecipeService.delete(id);
   }
 }
